feat(TASK-004): 입력 검증 구현 및 좌석 상한 우회 결함 수정
Refs: TASK-004

+++ backend/src/main/java/com/example/ticket_booking/service/SeatLimitExceededException.java
@@ -2,7 +2,7 @@ package com.example.ticket_booking.service;
 
 public class SeatLimitExceededException extends RuntimeException {
 
-  public SeatLimitExceededException(int totalSeats) {
+  public SeatLimitExceededException(long totalSeats) {
     super("좌석 총수는 5,000을 넘을 수 없습니다: " + totalSeats);
   }
 }
